Export Estimates

Allow the extraction of result estimates from a precomputed scc_result table.

exportEstimates(
  connectionDetails,
  connection,
  cdmDatabaseSchema,
  resultsTable,
  riskWindowsTable,
  negativeControlPairs = NULL,
  exposureIds = NULL,
  outcomeIds = NULL,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  databaseId,
  computeThreads = 1,
  analysisId = 1,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  resultsDatabaseSchema = NULL,
  firstOutcomeOnly = TRUE,
  resultExportPath = "scc_result",
  resultExportManager = getDefaultExportManager(resultExportPath, databaseId),
  controlType = "outcomes",
  diagnosticResults = NULL,
  diagnosticThresholds = getDefaultDiagnosticThresholds()
)

Arguments

connectionDetails

An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

connection

DatabaseConnector connection instance

cdmDatabaseSchema

Name of database schema that contains the OMOP CDM and vocabulary.

resultsTable

string - must be a permanent table

riskWindowsTable

string - must be a permanent table

negativeControlPairs

A list of vectors for pairs of negative control

exposureIds

A vector containing the drug_concept_ids or cohort_definition_ids of the exposures of interest. If empty, all exposures in the exposure table will be included.

outcomeIds

The condition_concept_ids or cohort_definition_ids of the outcomes of interest. If empty, all the outcomes in the outcome table will be included.

exposureDatabaseSchema

The name of the database schema that is the location where the exposure data used to define the exposure cohorts is available. If exposureTable = DRUG_ERA, exposureDatabaseSchema is not used by assumed to be cdmSchema. Requires read permissions to this database.

exposureTable

The tablename that contains the exposure cohorts. If exposureTable <> DRUG_ERA, then expectation is exposureTable has format of COHORT table: cohort_concept_id, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

outcomeDatabaseSchema

The name of the database schema that is the location where the data used to define the outcome cohorts is available. If exposureTable = CONDITION_ERA, exposureDatabaseSchema is not used by assumed to be cdmSchema. Requires read permissions to this database.

outcomeTable

The tablename that contains the outcome cohorts. If outcomeTable <> CONDITION_OCCURRENCE, then expectation is outcomeTable has format of COHORT table: COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.

databaseId

Unique identifier for database - required

computeThreads

Number of parallel threads for computing IRRs with exact confidence intervals.

analysisId

An integer unique to this analysis

tempEmulationSchema

Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

resultsDatabaseSchema

Schema to oputput results to. Ignored if resultsTable and riskWindowsTable are temporary.

firstOutcomeOnly

If TRUE, only use first occurrence of each condition concept id for each person.

resultExportPath

Folder where result files are exported

resultExportManager

ResultModelManager::ResultExportManager instance - customize this to implement an alternative mechanism for exporting results

controlType

Calibrate effect estimates with outcome (default) or exposure controls

diagnosticResults

Data frame of diagnostic results as returned by runSccDiagnostics

diagnosticThresholds

Named list of diagnostic thresholds. See getDefaultDiagnosticThresholds()